Tech: déplace les fonctions liées au feedback dans un module
